Remove Navigation.UserMenu
