+++ src/components/Frame/README.md
@@ -93,15 +93,6 @@ class FrameExample extends React.Component {
       },
     ];
 
-    const navigationUserMenuMarkup = (
-      <Navigation.UserMenu
-        actions={userMenuActions}
-        name="Dharma"
-        detail={storeName}
-        avatarInitials="D"
-      />
-    );
-
     const contextualSaveBarMarkup = isDirty ? (
       <ContextualSaveBar
         message="Unsaved changes"
@@ -157,7 +148,7 @@ class FrameExample extends React.Component {
     );
 
     const navigationMarkup = (
-      <Navigation location="/" userMenu={navigationUserMenuMarkup}>
+      <Navigation location="/">
         <Navigation.Section
           items={[
             {
