+++ src/components/Frame/tests/Frame.test.tsx
@@ -14,8 +14,6 @@ import {
   ContextualSaveBar as FrameContextualSavebar,
   Loading as FrameLoading,
 } from '../components';
-import TopBar, {UserMenuProvider} from '../../TopBar';
-import Navigation from '../../Navigation';
 
 window.matchMedia =
   window.matchMedia ||
@@ -218,27 +216,5 @@ describe('<Frame />', () => {
       </Frame>,
     );
     expect(frame.find(FrameLoading).exists()).toBe(true);
-  });
-
-  describe('<UserMenuProvider />', () => {
-    it('renders', () => {
-      const frame = mountWithAppProvider(<Frame />);
-      expect(frame.find(UserMenuProvider).exists()).toBe(true);
-    });
-
-    it('receives a mobileView boolean', () => {
-      const frame = mountWithAppProvider(<Frame />);
-      expect(frame.find(UserMenuProvider).prop('mobileView')).toBe(false);
-    });
-
-    it('receives the given top bar and navigation as its children', () => {
-      const topBar = <TopBar />;
-      const navigation = <Navigation location="" />;
-      const frame = mountWithAppProvider(
-        <Frame topBar={topBar} navigation={navigation} />,
-      );
-      expect(frame.find(UserMenuProvider).contains(topBar)).toBe(true);
-      expect(frame.find(UserMenuProvider).contains(navigation)).toBe(true);
-    });
   });
 });
