--- src/components/Navigation/components/UserMenu/index.ts
@@ -1,5 +1,0 @@
-import UserMenu from './UserMenu';
-
-export {Props} from './UserMenu';
-
-export default UserMenu;
