+++ src/components/Navigation/components/index.ts
@@ -1,4 +1,3 @@
 export {default as Section, Props as SectionProps} from './Section';
 export {default as Item, Props as ItemProps} from './Item';
 export {default as Message, Props as MessageProps} from './Message';
-export {default as UserMenu, Props as UserMenuProps} from './UserMenu';
