+++ src/components/TopBar/components/UserMenu/UserMenu.scss
