--- src/components/TopBar/components/UserMenu/components/UserMenu/index.ts
@@ -1,1 +1,0 @@
-export {default, Props} from './UserMenu';
