--- src/components/TopBar/components/UserMenu/components/index.ts
@@ -1,1 +1,0 @@
-export {default as UserMenu, Props as UserMenuProps} from './UserMenu';
