--- src/components/TopBar/components/UserMenu/context/Consumer.ts
@@ -1,6 +1,0 @@
-// Variable prefixed with underscore to prevent unused variable error.
-// More info here: https://github.com/Shopify/polaris-react/pull/938
-import * as _React from 'react';
-import UserMenuContext from './context';
-
-export default UserMenuContext.Consumer;
