--- src/components/TopBar/components/UserMenu/context/context.ts
@@ -1,10 +1,0 @@
-import React from 'react';
-import {UserMenuProps} from '../components';
-
-export interface UserMenuContextTypes {
-  mobileView?: boolean;
-  mobileUserMenuProps?: UserMenuProps;
-  setMobileUserMenuProps?(props: UserMenuProps): void;
-}
-
-export default React.createContext<UserMenuContextTypes>({});
