--- src/components/TopBar/components/UserMenu/context/index.ts
@@ -1,4 +1,0 @@
-export {default as Provider} from './Provider';
-export {default as Consumer} from './Consumer';
-export {default as Modifier} from './Modifier';
-export {default as UserMenuContext, UserMenuContextTypes} from './context';
