+++ src/components/TopBar/components/UserMenu/index.ts
@@ -1,5 +1,4 @@
 import UserMenu from './UserMenu';
 
-export {UserMenuProps as Props} from './components';
-export {Provider, Modifier} from './context';
+export {Props} from './UserMenu';
 export default UserMenu;
