+++ src/components/TopBar/components/index.ts
@@ -1,10 +1,5 @@
 export {default as Search, Props as SearchProps} from './Search';
 export {default as SearchField, Props as SearchFieldProps} from './SearchField';
-export {
-  default as UserMenu,
-  Props as UserProps,
-  Provider as UserMenuProvider,
-  Modifier as UserMenuModifier,
-} from './UserMenu';
+export {default as UserMenu, Props as UserProps} from './UserMenu';
 export {default as Menu, Props as MenuProps} from './Menu';
 export {MessageProps} from './Menu/components';
