+++ src/components/TopBar/index.ts
@@ -1,11 +1,5 @@
 import TopBar from './TopBar';
 
 export {Props} from './TopBar';
-export {
-  UserProps,
-  SearchFieldProps,
-  MessageProps,
-  UserMenuModifier,
-  UserMenuProvider,
-} from './components';
+export {UserProps, SearchFieldProps, MessageProps} from './components';
 export default TopBar;
